requirements.txt
beautifulsoup4==4.6.0
matplotlib==2.1.2
numpy==1.14.0
python-dateutil==2.6.0
pytz==2017.3
